% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ande4.R
\name{andersson}
\alias{andersson}
\alias{andersson.default}
\alias{coef.andersson}
\alias{plot.andersson}
\alias{predict.andersson}
\title{R optmisation of Andersson four-parameter particle size distribution model}
\usage{
andersson(data = NULL, D = NULL, fr = NULL, p = NULL, group = NULL,
  a = 1, b = 1, df1 = 1, do = 1)

\method{andersson}{default}(data = NULL, D = NULL, fr = NULL, p = NULL,
  group = NULL, a = 1, b = 1, df1 = 1, do = 1)

\method{plot}{andersson}(x, main = "Particle Size Distribution Function",
  xlab = "Particle Diameter", ylab = "Fraction", ...)

\method{coef}{andersson}(object, ...)

\method{predict}{andersson}(object, D, ...)
}
\arguments{
\item{data}{dataframe. It can contain data with column names of "time" and "I"}

\item{D}{character. The corresponding diameter in the data.}

\item{fr}{character. The corresponding fraction of the particles.}

\item{p}{porosity[g/m^3]}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{a}{an initial value of 'a' parameter to be optimised}

\item{b}{an initial value of 'b' parameter to be optimised}

\item{df1}{an initial value of 'df1' parameter to be optimised}

\item{do}{an initial value of 'do' parameter to be optimised}

\item{x}{a return object of the function.}

\item{main}{Title of the plot}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{...}{Any other graphical parameter}

\item{object}{Model output object}
}
\value{
\itemize{
\item{PD:} {  predicted fraction by mass of particles passing a particular diameter}
\item{sand:} { percentenge of sand}
\item{clay:} { percentage of clay}
\item{silt:} { percentage of silt }
\item{a:} { optimised a parameter }
\item{b:} { optimised b parameter }
\item{df1:} { optimised df1 parameter }
\item{do:} { optimised do parameter }
}
}
\description{
The function optimises 4 parameters of Andersson (1990) particle size distribution function
}
\examples{
data=read.csv(system.file("ext","sys","soil2.csv",package="vadose"))
single<- subset(data, ID=="30B20_1")
mod=andersson (data=single,p="sand",D="D",fr="Sand.")
plot(mod)
\dontrun{
#group simulation
mod2=andersson (data=data,D="D",fr="FractionSand",group="ID")
mod2$coef
}
#generic function
}
\author{
George Owusu
}
\references{
Andersson, S. (1990). Markfysikaliska undersokningar i odlad jord XXVI.
Om mineraljordens och mullens rumsutfyllande egenskaper. En teoretisk studie.
(In Swedish) (pp. 7). Uppsala: Swedish University of agricultural sciences.
}
\seealso{
\code{\link{zhuang3}}, \code{\link{zhuang4}},\code{\link{lass3}},\code{\link{logarithmic}},
\code{\link{logistic2}}, \code{\link{logistic3}},\code{\link{fredlund3}},\code{\link{fredlund4}},
 \code{\link{jaky}},\code{\link{andersson}},\code{\link{gompertz}},
\code{\link{haverkamp}}
}

