% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BEST.R
\name{BEST}
\alias{BEST}
\alias{BEST.default}
\alias{BESTlass}
\alias{BESTyb}
\alias{cal.BEST}
\alias{coef.BEST}
\alias{gof.BEST}
\alias{group.BEST}
\alias{plot.BEST}
\alias{predict.BEST}
\title{R estimation of Beerkan estimation of soil transfer (BEST)}
\usage{
BEST(data = NULL, time, I, S = 0.1, n = NULL, m = "b", y = 0.75,
  b = 0.6, r = 75, pb = 1.2, tho = 0.169, thr = 0, PSD = NULL,
  h = NULL, theta = NULL, steady = 3, init = 5, ths = NULL)

\method{BEST}{default}(data = NULL, time, I, S = 0.1, n = NULL, m = "b",
  y = 0.75, b = 0.6, r = 75, pb = 1.2, tho = 0.169, thr = 0,
  PSD = NULL, h = NULL, theta = NULL, steady = 3, init = 5,
  ths = NULL)

BESTlass(data = NULL, time, I, S = 0.1, n = NULL, m = "b", y = 0.75,
  b = 0.6, r = 75, pb = 1.2, tho = 0.169, thr = 0, PSD = NULL,
  h = NULL, theta = NULL, steady = 3, init = 5, ths = NULL)

BESTyb(data = NULL, time, I, S = 0.1, n = NULL, m = "b", y = 0.75,
  b = 0.6, r = 75, pb = 1.2, tho = 0.169, thr = 0, PSD = NULL,
  h = NULL, theta = NULL, steady = 3, init = 5, ths = NULL)

group.BEST(data = NULL, time, I, S = 0.1, n = NULL, m = "b", y = 0.75,
  b = 0.6, r = 75, pb = 1.2, tho = 0.169, thr = 0, PSD = NULL,
  h = NULL, theta = NULL, steady = 3, init = 5, ths = NULL, group,
  plot = TRUE, layout = c(2, 2), hlog = NULL, klog = NULL,
  opar = par(mar = c(2, 2, 1.8, 2)))

cal.BEST(data = NULL, time, I, S = 0.1, n = NULL, m = "b", y = c(0.6,
  0.8, 0.05), b = c(0.1, 0.9, 0.1), r = 75, pb = c(1, 1.8, 0.1),
  tho = c(0.01, 0.8, 0.05), thr = c(0, 0.2, 0.05), PSD = NULL, h = NULL,
  theta = NULL, steady = c(2, 10, 1), init = c(5, 7, 1), ths = NULL,
  model = "all", group = NULL)

gof.BEST(x = NULL, obs = NULL, est = NULL, P = NULL)

\method{coef}{BEST}(object, ...)

\method{predict}{BEST}(object, h = NULL, ...)

\method{plot}{BEST}(x, main = NULL, xlab = "Water Content", ylab = "",
  ylab2 = "", hlog = NULL, klog = NULL, col = c("blue", "red",
  "darkgreen"), units = c("s", "mm"), mfrow = c(2, 2), type = "all",
  legend = TRUE, opar = par(mar = c(2, 2, 1.5, 2)), ...)
}
\arguments{
\item{data}{dataframe. It can contain data with column names of "time" and "I"}

\item{time}{character or numeric. The name of time variable in the dataframe.
If the "data" parameter contains "time", this will be ignored.
The unit must be in seconds.}

\item{I}{character or numeric. The name of cumulative infiltration variable
in the dataframe. If the "data" parameter contains "I", this will be ignored.
The unit must be in millimetres [mm].}

\item{S}{Initial value of soil sorptivity [L T^(-1/2)]. The default is 0.1.
This will be optimise}

\item{n}{numeric. A shaping parameter for water retention curve. This can be calibrated.
If the parameter "PSD" is not set NULL \code{\link{lass3}} will be automatically used to estimate it.}

\item{m}{character. The water retention curve condition.
It takes either "b" for Burdine condition or "m" for Mualem condition.}

\item{y}{numeric. coefficient of BEST A parameter, commonly set at 0.75[-]
This can be calibrated.}

\item{b}{numeric. coefficient of BEST B parameter, commonly set at 0.6[-].
This can be calibrated.}

\item{r}{numeric. Radius of infiltrometer ring [mm]. The default is 75 mm}

\item{pb}{numeric. Bulk density[g cm^-3]. This can be calibrated.}

\item{tho}{numeric. initial volumetric soil water content. This can be calibrated.}

\item{thr}{numeric. Residual volumetric soil water content. This can be calibrated.}

\item{PSD}{dataframe. Particle Size Distribution. A dataframe of two columns.
The first column should be labeled "D" in the range of 0.001 to 2 mm.
The second column should be named as "fr" i.e.' the fraction of Diameter.
The range should be 0-1.}

\item{h}{list. The range of metric head desired or measured [mm].}

\item{theta}{The  measured soil water content[m3m-3] at the modelled pressure levels.
It is used to check the fitness of the model.}

\item{steady}{The number of data points that should be used for steady state.
This can be observed during infiltration measurement}

\item{init}{numeric. The number of data points that should be initially used for optimisation of
Best equation.}

\item{ths}{numeric. The saturated soil water content [m3/m3].}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{plot}{wheter a plot should be performed.}

\item{layout}{plot layout}

\item{hlog}{TRUE or FALSE. Whether metric potential should be log transformed}

\item{klog}{TRUE or FALSE. Whether hydraulic conductivity should be log transformed}

\item{model}{A calibration model parameter. It takes "all" for \code{{BEST}} "Lassabatere" for
\code{{BESTlass}} and "Yilmaz" for \code{{BESTyb}}, " Bagarello" \code{{BESTyb}}}

\item{x}{a return object of the function.}

\item{obs}{Observed data}

\item{est}{Estimated or modelled data}

\item{P}{Number of Parameters}

\item{object}{Model output object}

\item{...}{Any other graphical parameter}

\item{main}{Title of the plot}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{ylab2}{y label of the second plot (K)}

\item{col}{Color of the plot}

\item{units}{Units of the plot}

\item{mfrow}{The graphical layout of the plots}

\item{type}{The type of plot. It takes  "all" or "h"or "hk" or "k"}

\item{legend}{A legend of the plot}
}
\value{
The return paramaters can be assessed with \code{coef.BEST}.
The predicted soil water can be assesed with \code{predict.BEST}
In all the function returns BEST parametrs A, B, C, Ks, S, slope, intercept, mod_theta, hg
Beacuse \code{BEST} combines \code{BESTlass}  and \code{BESTyb} 'l', 'y',
and 'b' can be used to access Lassabatere etal(2006), Yilmaz et al(2010)
and Bagarello et al (2014) parameters, respectively.
For  the definitions of the output of PSD see \code{\link{lass3}}.
For the definitions of the output of goodness of fit tests see \code{\link{gof}}
\itemize{
\item{A:} {  constant[L^-1]}
\item{B:} { constant[L^-1]}
\item{C: } { constant[L^-1]}
\item{Ks,Ksl,Ksy,Ksb:} { saturated hydraulic conductivity for varoius algorithms [LT^-1] }
\item{S,Sl,Sy,Sb:} { Sorptivity for varoius algorithms [LT^-0.5]}
\item{slope,slopel,slopey:} { slope of steady state data for varoius algorithms}
\item{intercept,interceptl,intercepty:} { intercept of steady state data for varoius algorithms}
\item{mod_theta,mod_thetal,mod_thetay,mod_thetab:} { The estimated soil water content at
different metric potentionals for varoius algorithms }
\item{theta:} { The observed soil water content at different metric potentionals  }
\item{K,Kl,Ky,Kb:} { The unsaturated hydraulic conductivity [LT^-1] }
\item{Kr,Krl,Kry,Krb:} { The ratio of K and Ks }
\item{hg,hgl,hgy,hgb:} { bubbling capillary pressure for varoius algorithms [L] }
\item{init:} { the number of data points that should be initially used for
optimisation of Best equation [T]}
\item{q:} { infiltration rates [LT^-1] }
}
}
\description{
These functions computes the various versions of BEST.
\code{BEST} computes all the 3 versions of BEST:Lassabatere etal(2006), Yilmaz et al(2010) and Bagarello et al (2014).
\code{BESTlass} implements Lassabatere algorithm
\code{BESTyb} implements Lassabatere algorithms of Yilmaz and Bagarello et al. et al.
}
\details{
The function can perform normal BEST compution with \code{BEST}.
}
\examples{
\dontrun{
psd=read.csv(system.file("ext","sys","psd.csv",package="vadose"))
hf=read.csv(system.file("ext","sys","h.csv",package="vadose"))
data=read.csv(system.file("ext","sys","exampleBEST.csv",package="vadose"))
mod1<-BEST(data=data,time="time",I="I",h=hf$h,PSD=psd)

pred=predict(mod1)
coefficients=coef(mod1)

# with measured theta and h
mod<-BEST(data=data,time="time",I="I",h=hf$h,theta=mod1$mod_thetab,PSD=psd,y=0.7,b=0.6)
print(gof.BEST(mod))
plot(mod,hlog="yes",klog=TRUE)
par(mfrow=c(1,2))
plot(mod,type="h")
plot(mod,type="psd")
gof1=gof.BEST(mod)
print(gof1)

#calibration
cal.mod<-cal.BEST(data=data,time="time",I="I",h=hf$h,theta=mod1$mod_thetab,PSD=psd,steady=3,
init=5,thr=0,ths=NULL,tho=0.162)
#calibrate all, set the model to 'all' Lassabatere etal(2006),
Yilmaz et al(2010) and Bagarello et al (2014)
cal.modall<-cal.BEST(data=data,time="time",I="I",h=hf$h,theta=mod1$mod_thetab,PSD=psd,steady=3,
init=5,thr=0,ths=NULL,tho=0.162,model="all")
#get the paramters
parameters=coef.BEST(cal.modall)
S=parameters$S
A=parameters$A

}
}
\author{
George Owusu
}
\references{
\itemize{
\item{}{Lassabatere, L., Angulo-Jaramillo, R., Soria Ugalde, J. M., Cuenca, R.,
Braud, I., & Haverkamp, R. (2006). Beerkan Estimation of Soil Transfer
Parameters through Infiltration Experiments-BEST.
Soil Sci. Soc. Am. J., 70(2), 521-532. doi: 10.2136/sssaj2005.0026}
\item{}{Yilmaz, D., Lassabatere, L., Angulo-Jaramillo, R., Deneele, D., & Legret, M. (2010).
Hydrodynamic Characterization of Basic Oxygen Furnace Slag through an
Adapted BEST Method. Vadose Zone J., 9(1), 107-116. doi: 10.2136/vzj2009.0039}
 \item{}{Aiello, R., Bagarello, V., Barbagallo, S., Consoli, S., Di Prima, S., Giordano,
 G., & Iovino, M. (2014). An assessment of the Beerkan method for determining the
 hydraulic properties of a sandy loam soil. Geoderma, 235-236(0), 300-307.
 doi: http://dx.doi.org/10.1016/j.geoderma.2014.07.024}
\item{}{Bagarello, V., Di Prima, S., & Iovino, M. (2014).
Comparing Alternative Algorithms to Analyze the Beerkan Infiltration Experiment.
Soil Sci. Soc. Am. J., 78(3), 724-736. doi: 10.2136/sssaj2013.06.0231}
}
}

