% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brutsaert.R
\name{brutsaert}
\alias{brutsaert}
\alias{brutsaert.default}
\alias{coef.brutsaert}
\alias{plot.brutsaert}
\alias{predict.brutsaert}
\alias{print.brutsaert}
\alias{summary.brutsaert}
\title{Brutsaert infiltration parameter optmisation in R}
\usage{
brutsaert(data = NULL, time, I, S = 0.1, Ks = 0.1, B = 1, para = 2,
  group = NULL)

\method{brutsaert}{default}(data = NULL, time, I, S = 0.1, Ks = 0.1,
  B = 1, para = 2, group = NULL)

\method{predict}{brutsaert}(object, time = NULL, ...)

\method{plot}{brutsaert}(x, xlab = "Time(Minutes)",
  ylab = "Cumulative (mm)", main = NULL, layout = NULL, ...)

\method{summary}{brutsaert}(object)

\method{print}{brutsaert}(x, ...)

\method{coef}{brutsaert}(object, ...)
}
\arguments{
\item{data}{dataframe. It can contain data with column names of "time" and "I"}

\item{time}{character or numeric. The name of time variable in the dataframe.
If the "data" parameter contains "time", this will be ignored.
The unit must be in seconds.}

\item{I}{cumulative infiltration (I) or infiltration rate (i) [mm]}

\item{S}{numeric. sorptivity parameter}

\item{Ks}{saturated hydraulic conductivity}

\item{B}{brutsaert parameter}

\item{para}{numeric. Whether 2 or 3 parameter brutsaert model should be implemented.}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{object}{Model output object}

\item{...}{Any other graphical parameter}

\item{x}{a return object of the function.}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{main}{Title of the plot}

\item{layout}{plot layout}
}
\value{
\itemize{
\item{Ks:} {  optimised saturated hydraulic conductivity [L/T]}
\item{S:} { optimised sorptivity [LT^-0.5]}
\item{B:} { optmised B parameter}
\item{output:} { output of the group simulation}
}
}
\description{
This function optimises Brutsaert, W. (1977)
cumulative infiltration (I): B, Ks and S.
}
\examples{
data=read.csv(system.file("ext","sys","exampleBEST.csv",package="vadose"))
brutsaert1<-brutsaert(data=data,time="time",I="I")
#print(gof(brutsaert1))
plot(brutsaert1)
predict(brutsaert1)
coef(brutsaert1)
#gof(brutsaert1)

#group simulation
data=read.csv(system.file("ext","sys","infiltration2.csv",package="vadose"))
assin_breko<- subset(data, ID=="41A20_1")
brutsaertg<-brutsaert(data=data,time="minutes",I="CumInfil",group="ID")
coef(brutsaertg)
}
\author{
George Owusu
}
\references{
Brutsaert, W. (1977). Vertical infiltration in dry soil. Water Resour. Res., 13, 363-368.
}

