% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fred3.R
\name{fredlund3}
\alias{coef.fredlund3}
\alias{fredlund3}
\alias{fredlund3.default}
\alias{plot.fredlund3}
\alias{predict.fredlund3}
\title{R optmisation of Fredlund three-parameter particle size distribution function}
\usage{
fredlund3(data = NULL, D = NULL, fr = NULL, p = NULL, group = NULL,
  a = 1, m = 1, n = 1)

\method{fredlund3}{default}(data = NULL, D = NULL, fr = NULL, p = NULL,
  group = NULL, a = 1, m = 1, n = 1)

\method{plot}{fredlund3}(x, main = "Particle Size Distribution Function",
  xlab = "Particle Diameter", ylab = "Fraction", ...)

\method{coef}{fredlund3}(object, ...)

\method{predict}{fredlund3}(object, D, ...)
}
\arguments{
\item{data}{dataframe. It can contain data with column names of "time" and "I"}

\item{D}{character. The corresponding diameter in the data.}

\item{fr}{character. The corresponding fraction of the particles.}

\item{p}{porosity[g/m^3]}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{a}{an initial value of 'a' parameter to be optimised}

\item{m}{an initial value of 'm' parameter to be optimised}

\item{n}{an initial value of 'n' parameter to be optimised}

\item{x}{a return object of the function.}

\item{main}{Title of the plot}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{...}{Any other graphical parameter}

\item{object}{Model output object}
}
\value{
\itemize{
\item{PD:} {  predicted fraction by mass of particles passing a particular diameter}
\item{sand:} { percentenge of sand}
\item{clay:} { percentage of clay}
\item{silt:} { percentage of silt }
\item{a:} { optimised a parameter }
\item{m:} { optimised m parameter }
\item{n:} { optimised n parameter }
}
}
\description{
The function optimises 3 parameters of Fredlund et al (2000) particle size distribution function
}
\examples{
data=read.csv(system.file("ext","sys","soil2.csv",package="vadose"))
single<- subset(data, ID=="30B20_1")
mod=fredlund3 (data=single,p="sand",D="D",fr="Sand.")
plot(mod)

#group simulation
mod2=fredlund3 (data=data,D="D",fr="FractionSand",group="ID")
mod2$coef
#generic function
}
\author{
George Owusu
}
\references{
Fredlund, M. D., Fredlund, D. G., & Wilson, G. W. (2000).
An equation to represent grain-size distribution. Can Geotech J, 65, 638-648.
}
\seealso{
\code{\link{zhuang3}}, \code{\link{zhuang4}},\code{\link{lass3}},\code{\link{logarithmic}},
\code{\link{logistic2}}, \code{\link{logistic3}},\code{\link{logistic4}},\code{\link{fredlund4}},
 \code{\link{jaky}},\code{\link{andersson}},\code{\link{gompertz}},
\code{\link{haverkamp}}

Other Fredlund functions: \code{\link{coef.fredlund4}},
  \code{\link{fredlund4}}, \code{\link{fredlund4.default}},
  \code{\link{plot.fredlund4}},
  \code{\link{predict.fredlund4}}
}

