% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred4.R
\name{fredlund4}
\alias{coef.fredlund4}
\alias{fredlund4}
\alias{fredlund4.default}
\alias{plot.fredlund4}
\alias{predict.fredlund4}
\title{R optmisation of Fredlund et al (2000) four-parameter particle size distribution function}
\usage{
fredlund4(data = NULL, D = NULL, fr = NULL, p = NULL, group = NULL,
  a = 1, m = 1, n = 1, df = 1)

\method{fredlund4}{default}(data = NULL, D = NULL, fr = NULL, p = NULL,
  group = NULL, a = 1, m = 1, n = 1, df = 1)

\method{plot}{fredlund4}(x, main = "Particle Size Distribution Function",
  xlab = "Particle Diameter", ylab = "Fraction", ...)

\method{coef}{fredlund4}(object, ...)

\method{predict}{fredlund4}(object, D, ...)
}
\arguments{
\item{data}{dataframe. It can contain data with column names of "time" and "I"}

\item{D}{character. The corresponding diameter in the data.}

\item{fr}{character. The corresponding fraction of the particles.}

\item{p}{porosity[g/m^3]}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{a}{an initial value of 'a' parameter to be optimised}

\item{m}{character. The water retention curve condition. 
It takes either "b" for Burdine condition or "m" for Mualem condition.}

\item{n}{numeric. A shaping parameter for water retention curve. This can be calibrated.
If the parameter "PSD" is not set NULL \code{\link{lass3}} will be automatically used to estimate it.}

\item{df}{an initial value of 'df' parameter to be optimised}

\item{x}{a return object of the function.}

\item{main}{Title of the plot}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{...}{Any other graphical parameter}

\item{object}{Model output object}
}
\value{
\itemize{
\item{PD:} {  predicted fraction by mass of particles passing a particular diameter}
\item{sand:} { percentage of sand}
\item{clay:} { percentage of clay}
\item{silt:} { percentage of silt }
\item{a:} { optimised a parameter }
\item{m:} { optimised m parameter }
\item{n:} { optimised n parameter }
\item{df:} { optimised df parameter }
}
}
\description{
The function optimises 4 parameters of Fredlund et al (2000) particle size distribution function
}
\examples{
data=read.csv(system.file("ext","sys","soil2.csv",package="vadose"))
single<- subset(data, ID=="30B20_1")
mod=fredlund4 (data=single,p="sand",D="D",fr="Sand.")
plot(mod)

#group simulation
mod2=fredlund4 (data=data,D="D",fr="FractionSand",group="ID")
mod2$coef
#generic function
}
\author{
George Owusu
}
\references{
Fredlund, M. D., Fredlund, D. G., & Wilson, G. W. (2000). 
An equation to represent grain-size distribution. Can Geotech J, 65, 638-648.
}
\seealso{
\code{\link{zhuang3}}, \code{\link{zhuang4}},\code{\link{lass3}},\code{\link{logarithmic}},
\code{\link{logistic2}}, \code{\link{logistic3}},\code{\link{logistic4}},\code{\link{fredlund3}},
 \code{\link{jaky}},\code{\link{andersson}},\code{\link{gompertz}},
\code{\link{haverkamp}}

Other Fredlund functions: \code{\link{fredlund3}}
}

