% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lass3.R
\name{gof}
\alias{gof}
\title{Estimation of goodness of fit tests with model objects or comparing two models}
\usage{
gof(object = NULL, c = NULL, r = NULL, P = NULL)
}
\arguments{
\item{object}{Function return object}

\item{c}{comparing model, optional}

\item{r}{reference model}

\item{P}{The number of parameters}
}
\value{
\itemize{
\item{r2:} {  R-square [-]}
 \item{r2_adj:} { An adjusted R-square [-]}
\item{d:} { Index of agreement[-]}
\item{p_value:} { p value}
\item{sse:} { Sum of square error}
\item{sst:} { Sum of square total}
\item{AIC:} { AIC}
\item{BIC:} { BIC}
\item{RMSE:} { Root Mean Square Error}
\item{RMSE_adj:} { adjusted Root Mean Square Error}
\item{NRMSE} { Normalised Root Mean Square Error}
\item{NRMSE_adj:} { adjusted Normalised Root Mean Square Error}
\item{E:} { Nash Sutcliffe Coefficient}
\item{Cp:} { statistic (Mallows 1973)}
\item{F1:} { F-statistic}
\item{AAE:} { average absolute error}
\item{ARE:} { average relative error}
\item{MBE:} { Mean Bias  error}
\item{MPE:} { Mean Percentage error}
}
}
\description{
Estimation of goodness of fit tests with model objects or comparing two models
}
\examples{
data=read.csv(system.file("ext","sys","soil2.csv",package="vadose"))
single<- subset(data, ID=="30B20_1")
mod=lass3 (data=single,p="sand",D="D",fr="Sand.")
fit=gof(mod)
r2=fit$r2
AIC=fit$AIC
print (fit)

data=read.csv(system.file("ext","sys","exampleBEST.csv",package="vadose"))
philip1<-philip(data=data,time="time",I="I")
sw1<-sw(data=data,time="time",I="I")
thisgof=gof(c=gof(sw1),r=gof(philip1))
}
\author{
George Owusu
}
\references{
Mallows, C. L. (1973). Some comments on Cp. Technometrics, 15, 661-675.
}

