% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gomp4.R
\name{gompertz}
\alias{coef.gompertz}
\alias{gompertz}
\alias{gompertz.default}
\alias{plot.gompertz}
\alias{predict.gompertz}
\title{R optmisation of Nemes et al (1999) four-parameter particle size distribution model}
\usage{
gompertz(data = NULL, D = NULL, fr = NULL, p = NULL, group = NULL,
  a = 1, b = 1, e = 1, z = 1)

\method{gompertz}{default}(data = NULL, D = NULL, fr = NULL, p = NULL,
  group = NULL, a = 1, b = 1, e = 1, z = 1)

\method{plot}{gompertz}(x, main = "Particle Size Distribution Function",
  xlab = "Particle Diameter", ylab = "Fraction", ...)

\method{coef}{gompertz}(object, ...)

\method{predict}{gompertz}(object, D, ...)
}
\arguments{
\item{data}{dataframe. It can contain data with column names of "time" and "I"}

\item{D}{character. The corresponding diameter in the data.}

\item{fr}{character. The corresponding fraction of the particles.}

\item{p}{porosity[g/m^3]}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{a}{an initial value of 'a' parameter to be optimised}

\item{b}{an initial value of 'b' parameter to be optimised}

\item{e}{an initial value of 'e' parameter to be optimised}

\item{z}{an initial value of 'z' parameter to be optimised}

\item{x}{a return object of the function.}

\item{main}{Title of the plot}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{...}{Any other graphical parameter}

\item{object}{Model output object}
}
\value{
\itemize{
\item{PD:} {  predicted fraction by mass of particles passing a particular diameter}
\item{sand:} { percentenge of sand}
\item{clay:} { percentage of clay}
\item{silt:} { percentage of silt }
\item{a:} { optimised a parameter }
\item{b:} { optimised b parameter }
\item{e:} { optimised e parameter }
\item{z:} { optimised z parameter }
}
}
\description{
The function optimises 4 parameters of Nemes et al (1999) particle size distribution function
}
\examples{
data=read.csv(system.file("ext","sys","soil2.csv",package="vadose"))
single<- subset(data, ID=="30B20_1")
mod=gompertz (data=single,p="sand",D="D",fr="Sand.")
plot(mod)
\dontrun{
#group simulation
mod2=gompertz (data=data,D="D",fr="FractionSand",group="ID")
mod2$coef
}
#generic function
}
\author{
George Owusu
}
\references{
Nemes, A., Wosten, J. H. M., Lilly, A., & Voshaar, J. H. O. (1999).
Evaluation of different procedures to interpolate particle-size distributions
to achieve compatibility within soil databases. Geoderma., 90, 187-202.
}
\seealso{
\code{\link{zhuang3}}, \code{\link{zhuang4}},\code{\link{lass3}},\code{\link{logarithmic}},
\code{\link{logistic2}}, \code{\link{logistic3}},\code{\link{fredlund3}},\code{\link{fredlund4}},
 \code{\link{jaky}},\code{\link{andersson}},\code{\link{gompertz}},
\code{\link{haverkamp}}
}

