% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenampt.R
\name{greenampt}
\alias{coef.greenampt}
\alias{greenampt}
\alias{greenampt.default}
\alias{plot.greenampt}
\alias{predict.greenampt}
\alias{print.greenampt}
\alias{summary.greenampt}
\title{Green and Ampt infiltration parameter optmisation in R}
\usage{
greenampt(data, time, I, Ks = 0.1, G = 0.1, group = NULL)

\method{greenampt}{default}(data, time, I, Ks = 0.1, G = 0.1,
  group = NULL)

\method{predict}{greenampt}(object, time = NULL, ...)

\method{plot}{greenampt}(x, xlab = "Time(Minutes)",
  ylab = "Cumulative (cm)", main = NULL, layout = NULL, ...)

\method{summary}{greenampt}(object, ...)

\method{print}{greenampt}(x, ...)

\method{coef}{greenampt}(object, ...)
}
\arguments{
\item{data}{dataframe. It can contain data with column names of "time" and "I"}

\item{time}{character or numeric. The name of time variable in the dataframe.
If the "data" parameter contains "time", this will be ignored.
The unit must be in seconds.}

\item{I}{cumulative infiltration (I) or infiltration rate (i) [mm]}

\item{Ks}{Hydraulic conductivity}

\item{G}{Green and Ampt parameter that is equivalent to Sorptivity}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{object}{Model output object}

\item{...}{Any other graphical parameter}

\item{x}{a return object of the function.}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{main}{Title of the plot}

\item{layout}{plot layout}
}
\value{
\itemize{
\item{Ks:} {  Hydraulic conductivity [L]}
\item{G:} { Green and Ampt parameter that is equivalent to Sorptivity [LT^0.5]}
\item{predict:}{predicted infiltration}
\item{output:} { output of the group simulation}
}
}
\description{
This function optimises Green and Ampt (1911)
 infiltration parameters: Ks and G. It also predicts infiltration.
}
\examples{
data=read.csv(system.file("ext","sys","exampleBEST.csv",package="vadose"))
greenampt1<-greenampt(data=data,time="time",I="I")
#print(gof(greenampt1))
#plot(greenampt1)
predict(greenampt1)
coef(greenampt1)

#group simulation
data=read.csv(system.file("ext","sys","infiltration2.csv",package="vadose"))
greenampt1g<-greenampt(data=data,time="minutes",I="CumInfil",group="ID")
coef(greenampt1g)
#generic function ######################################
}
\author{
George Owusu
}
\references{
Green, W. A., & Ampt, G. A., 1911..4,1-24. (1911). Studies on soil  physics:1.
The flow of air and water through soils. Journal of Agricultural Science, 4(1-24).
}

