% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/van2haverkamp.R
\name{haverkamp}
\alias{coef.haverkamp}
\alias{haverkamp}
\alias{haverkamp.default}
\alias{plot.haverkamp}
\alias{predict.haverkamp}
\title{R optmisation of a two-parameter van Genuchten particle size distribution function}
\usage{
haverkamp(data = NULL, D = NULL, fr = NULL, a = 1, b = 1, p = NULL,
  group = NULL)

\method{haverkamp}{default}(data = NULL, D = NULL, fr = NULL, a = 1,
  b = 1, p = NULL, group = NULL)

\method{plot}{haverkamp}(x, main = "Particle Size Distribution Function",
  xlab = "Particle Diameter", ylab = "Fraction", ...)

\method{coef}{haverkamp}(object, ...)

\method{predict}{haverkamp}(object, D, ...)
}
\arguments{
\item{data}{a dataframe of Particle Size Distribution. 
It can have column names of diameter of the particles "D" [mm], 
and fraction of the particles "fr" that ranges from 0-1.}

\item{D}{character. The corresponding diameter in the data.}

\item{fr}{character. The corresponding fraction of the particles.}

\item{a}{an initial value of 'a' parameter to be optimised}

\item{b}{an initial value of 'b' parameter to be optimised}

\item{p}{porosity[g/m^3]}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{x}{a return object of the function.}

\item{main}{Title of the plot}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{...}{Any other graphical parameter}

\item{object}{Model output object}
}
\value{
\itemize{
\item{PD:} {  predicted fraction by mass of particles passing a particular diameter}
\item{sand:} { percentage of sand}
\item{clay:} { percentage of clay}
\item{silt:} { percentage of silt }
\item{a:} { optimised a parameter }
\item{b:} { optimised b parameter }
}
}
\description{
The function optimises 2-parameter van Genuchten  particle size distribution function
of Haverkamp & Parlange (1986).
}
\examples{
data=read.csv(system.file("ext","sys","soil2.csv",package="vadose"))
single<- subset(data, ID=="30B20_1")
mod=haverkamp (data=single,p="sand",D="D",fr="Sand.")
plot(mod)
\dontrun{
#group simulation
mod2=haverkamp (data=data,D="D",fr="FractionSand",group="ID")
mod2$coef
}
}
\author{
George Owusu
}
\references{
Haverkamp, R., & Parlange, J. Y. (1986). Predicting the Water-Retention Curve from 
Particle-Size Distribution:1. Sandy Soils Without Organic Matter. Soil Sc, 142, 325-339.
}
\seealso{
\code{\link{zhuang3}}, \code{\link{zhuang4}},\code{\link{lass3}},\code{\link{logarithmic}},
\code{\link{logistic2}}, \code{\link{logistic3}},\code{\link{logistic4}},\code{\link{fredlund3}},
\code{\link{fredlund4}}, \code{\link{jaky}},\code{\link{andersson}},\code{\link{gompertz}}

Other van-Genuchten functions: \code{\link{zhuang3}},
  \code{\link{zhuang4}}
}

