% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horton.R
\name{horton}
\alias{coef.horton}
\alias{horton}
\alias{horton.default}
\alias{plot.horton}
\alias{predict.horton}
\alias{print.horton}
\alias{summary.horton}
\title{Horton infiltration parameter optmisation in R}
\usage{
horton(data = NULL, time, I, fc = 0.1, fo = 0.1, k = 0.1,
  group = NULL)

\method{horton}{default}(data = NULL, time, I, fc = 0.1, fo = 0.1,
  k = 0.1, group = NULL)

\method{predict}{horton}(object, time = NULL, ...)

\method{plot}{horton}(x, xlab = "Time(Minutes)", ylab = "Cumulative (cm)",
  main = NULL, layout = NULL, ...)

\method{summary}{horton}(object)

\method{print}{horton}(x, ...)

\method{coef}{horton}(object, ...)
}
\arguments{
\item{data}{dataframe. It can contain data with column names of "time" and "I"}

\item{time}{character or numeric. The name of time variable in the dataframe.
If the "data" parameter contains "time", this will be ignored.
The unit must be in seconds.}

\item{I}{cumulative infiltration (I) or infiltration rate (i) [mm]}

\item{fc}{final constant rate of infiltration at saturation}

\item{fo}{parameter for initial infiltration capacity}

\item{k}{a constant depending primarily upon soil and vegetation}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{object}{Model output object}

\item{...}{Any other graphical parameter}

\item{x}{a return object of the function.}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{main}{Title of the plot}

\item{layout}{plot layout}
}
\value{
\itemize{
\item{fc:} {  final constant rate of infiltration at saturation}
\item{fo:} { parameter for initial infiltration capacity}
\item{k:} { a constant depending primarily upon soil and vegetation}
\item{predict:}{predicted infiltration}
\item{output:} { output of the group simulation}
}
}
\description{
This function optimises Horton (1940)
 infiltration parameters: fo,fc, and k. It also predicts infiltration.
}
\examples{
data=read.csv(system.file("ext","sys","exampleBEST.csv",package="vadose"))
horton1<-horton(data=data,time="time",I="I")
#print(gof(horton1))
plot(horton1)
predict(horton1)
coef(horton1)

#group simulation
data=read.csv(system.file("ext","sys","infiltration2.csv",package="vadose"))
horton1g<-horton(data=data,time="minutes",I="CumInfil",group="ID")
coef(horton1g)
#generic function ######################################
}
\author{
George Owusu
}
\references{
Horton, R. E. (1940). An approach toward a physical 
interpretation of infiltration capacity Proc. Sci. Soc. Amer., 5, 399-417.
}

