% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/offinbest.R
\docType{data}
\name{htheta}
\alias{htheta}
\title{MeasureD water retention data}
\format{A data frame with 90 rows and 4 variables:
\describe{
  \item{ID}{Unique ID of the data}
  \item{TownName}{Town Name}
  \item{hbar}{Metric potentials in bar}
  \item{h}{Metric potential in mm}
  \item{theta}{gravimetric soil water content [-]}
}}
\description{
A dataset hydraulic heads and water content.
}
\author{
George Owusu
}

