% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaky1.R
\name{jaky}
\alias{coef.jaky}
\alias{jaky}
\alias{jaky.default}
\alias{plot.jaky}
\alias{predict.jaky}
\title{R optmisation of Jaky particle size distribution function}
\usage{
jaky(data = NULL, D = NULL, fr = NULL, j = 1, p = NULL,
  group = NULL)

\method{jaky}{default}(data = NULL, D = NULL, fr = NULL, j = 1,
  p = NULL, group = NULL)

\method{plot}{jaky}(x, main = "Particle Size Distribution Function",
  xlab = "Particle Diameter", ylab = "Fraction", ...)

\method{coef}{jaky}(object, ...)

\method{predict}{jaky}(object, D, ...)
}
\arguments{
\item{data}{dataframe. It can contain data with column names of "time" and "I"}

\item{D}{character. The corresponding diameter in the data.}

\item{fr}{character. The corresponding fraction of the particles.}

\item{j}{initial value of j parameter to be optimised}

\item{p}{porosity[g/m^3]}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{x}{a return object of the function.}

\item{main}{Title of the plot}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{...}{Any other graphical parameter}

\item{object}{Model output object}
}
\value{
\itemize{
\item{PD:} {  predicted fraction by mass of particles passing a particular diameter}
\item{D10: } { The diameter at which 10 per cent of the sample mass comprised of smaller particles }
\item{D60: } { The diameter at which 60 per cent of the sample mass comprised of smaller particles }
\item{sand:} { percentage of sand}
\item{clay:} { percentage of clay}
\item{silt:} { percentage of silt }
\item{j:} { optimised j parameter }
}
}
\description{
The function optimises parameters of Jaky(1944) particle size distribution function
}
\examples{
data=read.csv(system.file("ext","sys","soil2.csv",package="vadose"))
single<- subset(data, ID=="30B20_1")
mod=jaky (data=single,p="sand",D="D",fr="Sand.")
plot(mod)

#group simulation
mod2=jaky (data=data,D="D",fr="FractionSand",group="ID")
mod2$coef
#generic function

#input function ############################################
}
\author{
George Owusu
}
\references{
Jaky, J. (1944). Soil mechanics. (In Hungarian). Budapest: Egyetemi Nyomda.
}
\seealso{
\code{\link{zhuang3}}, \code{\link{zhuang4}},\code{\link{lass3}},\code{\link{logarithmic}},
\code{\link{logistic2}}, \code{\link{logistic3}},\code{\link{fredlund3}},\code{\link{fredlund4}},
 \code{\link{jaky}},\code{\link{andersson}},\code{\link{gompertz}},
\code{\link{haverkamp}}
}

