% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ko.R
\name{kostiakov}
\alias{kostiakov}
\alias{kostiakov.default}
\alias{plot.kostiakov}
\alias{predict.kostiakov}
\alias{summary.kostiakov}
\title{Kostiakov infiltration parameter optmisation in R}
\usage{
kostiakov(data = NULL, time, I, a = 0.1, k = 0.1, group = NULL)

\method{kostiakov}{default}(data = NULL, time, I, a = 0.1, k = 0.1,
  group = NULL)

\method{predict}{kostiakov}(object, time = NULL, ...)

\method{plot}{kostiakov}(x, xlab = "Time(Minutes)",
  ylab = "Cumulative (cm)", main = NULL, layout = NULL, ...)

\method{summary}{kostiakov}(object, ...)
}
\arguments{
\item{data}{dataframe. It can contain data with column names of "time" and "I"}

\item{time}{character or numeric. The name of time variable in the dataframe.
If the "data" parameter contains "time", this will be ignored.
The unit must be in seconds.}

\item{I}{cumulative infltration (I) or infiltration rate (i) [mm]}

\item{a}{constant}

\item{k}{constant}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{object}{Model output object}

\item{...}{Any other graphical parameter}

\item{x}{a return object of the function.}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{main}{Title of the plot}

\item{layout}{plot layout}
}
\value{
\itemize{
\item{a:} {  optimised constant}
\item{k:} { optimised constant}
\item{output:} { output of the group simulation}
}
}
\description{
This function optimises Kostiakov (1932)
cumulative infltration (I) parameters.
}
\examples{
data=read.csv(system.file("ext","sys","exampleBEST.csv",package="vadose"))
kostiakov1<-kostiakov(data=data,time="time",I="I")
#print(gof(kostiakov1))
plot(kostiakov1)
predict(kostiakov1)
coef(kostiakov1)
#gof(kostiakov1)

#group simulation
data=read.csv(system.file("ext","sys","infiltration2.csv",package="vadose"))
assin_breko<- subset(data, ID=="41A20_1")
kostiakovg<-kostiakov(data=data,time="minutes",I="CumInfil",group="ID")
coef(kostiakovg)
}
\author{
George Owusu
}
\references{
Kostiakov, A. N. (1932). On the dynamics of the coefficient of water-percolation in soils and on the
necessity for studying it from a dynamic point of view for purposes of amelioration. 6.
}

