% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ksat.R
\name{ksat}
\alias{ksat}
\alias{ksat.default}
\title{Estimating saturated hydraulic conductivity with water retention parameters}
\usage{
ksat(ths = NULL, thr = 0, alpha = NULL, hb = NULL, th33 = NULL,
  th1500 = NULL, C = NULL, f = NULL, para = NULL, n = NULL,
  model = "nvr")

\method{ksat}{default}(ths = NULL, thr = 0, alpha = NULL, hb = NULL,
  th33 = NULL, th1500 = NULL, C = NULL, f = NULL, para = NULL,
  n = NULL, model = "nvr")
}
\arguments{
\item{ths}{numeric. The saturated soil water content [m3/m3].}

\item{thr}{numeric. Residual volumetric soil water content. This can be calibrated.}

\item{alpha}{van Genuchten's  water retention parameter.
It is equal to 1/hb}

\item{hb}{the bubbling capillary pressure parameter of Brooks and Corey. It is equal to 1/apha.}

\item{th33}{The water content at field capacity [m3/m3] .i.e metric potential of 33 kPa}

\item{th1500}{The water content at wilting point [m3/m3] i.e metric potential of 1500 kPa}

\item{C}{clay percentage}

\item{f}{porosity}

\item{para}{a parameter for each model}

\item{n}{van Genuchten's n that is realted to pore-size distribution.}

\item{model}{model type. It can take  "Nasta" (default), "Mishra", "RAWLS", "rawls2006",
"Guarracino".}
}
\value{
\itemize{
\item{Ks:} {  saturated hydraulic conductivity}
\item{para:} { soil parameters}
}
}
\description{
This function estimates Ks based on 5 models that include (see references)
Guarracino (2007), Nasta et al (2013), Mishra et al(1990), Rawls et al(1982) and
Saxton & Rawls (2006). The model can also return tabulated Ks of Carsel and Parrish (1988)
when soil texture is given as a "model" parameters.
}
\examples{
data=read.csv(system.file("ext","sys","retentionVG.csv",package="vadose"))
mod<-vg(data=data,h="h",theta="theta",thr=0.1, ths=0.1, alp=0.1, n=1,Ks="nvr")
nvr=ksat(ths=mod$ths,hb=1/mod$alp,n=mod$n-1,model="Nasta")
print(nvr)
g=ksat(ths=mod$ths,alpha=mod$alp,model="Guarracino")
print(g$Ks)
soil=ksat(ths=mod$ths,alpha=mod$alp,model="Clay")
soil=ksat(para="soil",model="Clay")
mp=ksat(ths=mod$ths,alpha=mod$alp,model="Mishra")
print(mp)
r=ksat(ths=mod$ths,hb=1/mod$alp,n=mod$n-1,model="RAWLS",thr=mod$thr,f=mod$ths,para=86)
print(r)
sr=ksat(ths=mod$ths,th33=predict(mod,33),th1500=predict(mod,1500),model="rawls2006")
print(sr)
}
\references{
\itemize{
\item{}{Guarracino, L. (2007). Estimation of saturated hydraulic conductivity
Ks from the van Genuchten shape parameter. Water Resour. Res., 43.
doi: doi:10.1029/2006WR005766}
\item{}{Paolo Nasta, Jasper A. Vrugt, & Romano, N. (2013). Prediction of the
saturated hydraulic conductivity from Brooks and Corey's
water retention parameters. WATER RESOURCES RESEARCH, 49. doi: doi:10.1002/wrcr.20269}
\item{}{Mishra, S., & Parker, J. C. (1990). On the relation between
saturated conductivity and capillary retention characteristics.
Ground Water, 28, 775-777.}
\item{}{Rawls, W. J., Brakensiek, D. L., & Saxon, K. E. (1982).
Estimation of soil water properties. Transactions of the ASAE, 25, 1316-1320.}
\item{}{Saxton, K. E., & Rawls, W. J. (2006). Soil Water Characteristic Estimates by
Texture and Organic Matter for Hydrologic Solutions.
SOIL SCI. SOC. AM. J., 70. doi: 10.2136/sssaj2005.0117}
}
}

