% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lass3.R
\name{lass3}
\alias{coef.lass3}
\alias{lass3}
\alias{lass3.default}
\alias{plot.lass3}
\alias{predict.lass3}
\title{R Estimation of the Water Retention Shape Parameters from particle size distribution}
\usage{
lass3(data = NULL, D = NULL, fr = NULL, N = 0.1, M = 0.1, Dg = 0.1,
  p = NULL, S = NULL, C = NULL, ksat = NULL, group = NULL)

\method{lass3}{default}(data = NULL, D = NULL, fr = NULL, N = 0.1,
  M = 0.1, Dg = 0.1, p = NULL, S = NULL, C = NULL, ksat = NULL,
  group = NULL)

\method{plot}{lass3}(x, main = "Particle Size Distribution Function",
  xlab = "Particle Diameter", ylab = "Fraction", ...)

\method{coef}{lass3}(object, ...)

\method{predict}{lass3}(object, D, ...)
}
\arguments{
\item{data}{a dataframe of Particle Size Distribution. 
It can have column names of diameter of the particles "D" [mm], 
and fraction of the particles "fr" that ranges from 0-1.}

\item{D}{character. The corresponding diameter in the data.}

\item{fr}{character. The corresponding fraction of the particles.}

\item{N}{numeric. The initial value of the shape parameter.}

\item{M}{numeric. The initial value of the shape parameter.}

\item{Dg}{numeric. A scale parameter}

\item{p}{porosity[g/m^3]}

\item{S}{The percentage of sand in the soil. This is needed if there is no data on PSD.}

\item{C}{The percentage of clay in the soil. This is needed if there is no data on PSD}

\item{ksat}{Saturated hydraulic conductivity. If NULL, this will be internally  estimated with
\code{\link{ksat}}}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{x}{a return object of the function.}

\item{main}{Title of the plot}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{...}{Any other graphical parameter}

\item{object}{Model output object}
}
\value{
\itemize{
\item{PD:} {  predicted fraction by mass of particles passing a particular diameter}
\item{poresizeindex:} { pore size distribution index[-]}
\item{D10: } { The diameter at which 10 per cent of the sample mass comprised of smaller particles }
\item{D60: } { The diameter at which 60 per cent of the sample mass comprised of smaller particles }
\item{sand:} { percentage of sand}
\item{clay:} { percentage of clay}
\item{silt:} { percentage of silt }
\item{n:} { shape parameter of van Genuchten water retention alpha parameter }
}
}
\description{
The function implements Lassabatere et al (2006) particle size distribution function
for estimation of Water Retention curve. It also provides the parameters based on 
Minasny & McBratney (2007) studies.
}
\examples{
data=read.csv(system.file("ext","sys","soil2.csv",package="vadose"))
single<- subset(data, ID=="30B20_1")
mod=lass3 (data=single,p="sand",D="D",fr="Sand.")
plot(mod)

#group simulation
mod2=lass3 (data=data,D="D",fr="FractionSand",group="ID",ksat=TRUE)
mod2$coef
#generic function
}
\author{
George Owusu
}
\references{
\itemize{
\item{}{Lassabatere, L., Angulo-Jaramillo, R., Soria Ugalde, J. M., Cuenca, R., Braud, I., & Haverkamp, R. (2006). 
Beerkan Estimation of Soil Transfer Parameters through Infiltration 
Experiments-BEST. Soil Sci. Soc. Am. J., 70(2), 521-532. 
doi: 10.2136/sssaj2005.0026}
\item{}{Minasny, B., & McBratney, A. B. (2007). Estimating the Water Retention Shape Parameter from 
Sand and Clay Content Soil Sci. Soc. Am. J., 71(4), 1105-1110. doi: 10.2136/sssaj2006.0298N}
}
}
\seealso{
\code{\link{zhuang3}}, \code{\link{zhuang4}},\code{\link{logarithmic}},
\code{\link{logistic2}}, \code{\link{logistic3}},\code{\link{fredlund3}},\code{\link{fredlund4}},
 \code{\link{jaky}},\code{\link{andersson}},\code{\link{gompertz}},
\code{\link{haverkamp}}
}

