% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logl2.R
\name{logarithmic}
\alias{coef.logarithmic}
\alias{logarithmic}
\alias{logarithmic.default}
\alias{plot.logarithmic}
\alias{predict.logarithmic}
\title{R optmisation of a two-parameter logarithmic particle size distribution function}
\usage{
logarithmic(data = NULL, D = NULL, fr = NULL, a = 0.1, b = 0.1,
  p = NULL, group = NULL)

\method{logarithmic}{default}(data = NULL, D = NULL, fr = NULL, a = 0.1,
  b = 0.1, p = NULL, group = NULL)

\method{plot}{logarithmic}(x, main = "Particle Size Distribution Function",
  xlab = "Particle Diameter", ylab = "Fraction", ...)

\method{coef}{logarithmic}(object, ...)

\method{predict}{logarithmic}(object, D, ...)
}
\arguments{
\item{data}{a dataframe of Particle Size Distribution.
It can have column names of diameter of the particles "D" [mm],
and fraction of the particles "fr" that ranges from 0-1.}

\item{D}{character. The corresponding diameter in the data.}

\item{fr}{character. The corresponding fraction of the particles.}

\item{a}{an initial value of 'a' parameter to be optimised}

\item{b}{an initial value of 'b' parameter to be optimised}

\item{p}{porosity[g/m^3]}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{x}{a return object of the function.}

\item{main}{Title of the plot}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{...}{Any other graphical parameter}

\item{object}{Model output object}
}
\value{
\itemize{
\item{PD:} {  predicted fraction by mass of particles passing a particular diameter}
\item{sand:} { percentenge of sand}
\item{clay:} { percentage of clay}
\item{silt:} { percentage of silt }
\item{a:} { optimised a parameter }
\item{b:} { optimised b parameter }
}
}
\description{
The function optimises 2-parameter logarithmic  particle size distribution function
of Zhuang et al(2001)
}
\examples{
data=read.csv(system.file("ext","sys","soil2.csv",package="vadose"))
single<- subset(data, ID=="30B20_1")
mod=logarithmic (data=single,p="sand",D="D",fr="Sand.")
plot(mod)
\dontrun{
#group simulation
mod2=logarithmic (data=data,D="D",fr="FractionSand",group="ID")
mod2$coef
}
}
\author{
George Owusu
}
\references{
Zhuang, J., Jin, Y., & Miyazaki, T. (2001). Estimating water retention characteristic from
soil particle-size distribution using a non-similar media concept. Soil Sci., 166, 308-321.
}
\seealso{
\code{\link{zhuang3}}, \code{\link{zhuang4}},\code{\link{lass3}},\code{\link{logarithmic}},
\code{\link{logistic2}}, \code{\link{logistic3}},\code{\link{logistic4}},\code{\link{fredlund3}},
\code{\link{fredlund4}}, \code{\link{jaky}},\code{\link{andersson}},\code{\link{gompertz}},
\code{\link{haverkamp}}
}

