% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logis4.R
\name{logistic4}
\alias{coef.logistic4}
\alias{logistic4}
\alias{logistic4.default}
\alias{plot.logistic4}
\alias{predict.logistic4}
\title{R optmisation of a four-parameter logistic particle size distribution function}
\usage{
logistic4(data = NULL, D = NULL, fr = NULL, p = NULL, group = NULL,
  a = 1, b = 1, e = 1, z = 1)

\method{logistic4}{default}(data = NULL, D = NULL, fr = NULL, p = NULL,
  group = NULL, a = 1, b = 1, e = 1, z = 1)

\method{plot}{logistic4}(x, main = "Particle Size Distribution Function",
  xlab = "Particle Diameter", ylab = "Fraction", ...)

\method{coef}{logistic4}(object, ...)

\method{predict}{logistic4}(object, D, ...)
}
\arguments{
\item{data}{a dataframe of Particle Size Distribution. 
It can have column names of diameter of the particles "D" [mm], 
and fraction of the particles "fr" that ranges from 0-1.}

\item{D}{character. The corresponding diameter in the data.}

\item{fr}{character. The corresponding fraction of the particles.}

\item{p}{porosity[g/m^3]}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{a}{an initial value of 'a' parameter to be optimised}

\item{b}{an initial value of 'b' parameter to be optimised}

\item{e}{an initial value of 'e' parameter to be optimised}

\item{z}{an initial value of 'z' parameter to be optimised}

\item{x}{a return object of the function.}

\item{main}{Title of the plot}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{...}{Any other graphical parameter}

\item{object}{Model output object}
}
\value{
\itemize{
\item{PD:} {  predicted fraction by mass of particles passing a particular diameter}
\item{sand:} { percentage of sand}
\item{clay:} { percentage of clay}
\item{silt:} { percentage of silt }
\item{a:} { optimised a parameter }
\item{b:} { optimised b parameter }
\item{e:} { optimised e parameter }
\item{z:} { optimised z parameter }
}
}
\description{
The function optimises 4-parameter logistic  particle size distribution function
of Skaggs et al(2001)
}
\examples{
data=read.csv(system.file("ext","sys","soil2.csv",package="vadose"))
single<- subset(data, ID=="30B20_1")
mod=logistic4 (data=single,p="sand",D="D",fr="Sand.")
plot(mod)
\dontrun{
#group simulation
mod2=logistic4 (data=data,D="D",fr="FractionSand",group="ID")
mod2$coef
}
}
\author{
George Owusu
}
\references{
Skaggs, T. H., Arya, L. M., Shouse, P. J., & Mohanty, B. P. (2001). 
Estimating Particle-Size Distribution from Limited Soil Texture Data. 
Soil Sci Soc Am J., 65, 1038-1044.
}
\seealso{
\code{\link{zhuang3}}, \code{\link{zhuang4}},\code{\link{lass3}},\code{\link{logarithmic}},
\code{\link{logistic2}}, \code{\link{logistic3}},\code{\link{fredlund3}},\code{\link{fredlund4}},
 \code{\link{jaky}},\code{\link{andersson}},\code{\link{gompertz}},
\code{\link{haverkamp}}

Other Logistic functions: \code{\link{logistic2}},
  \code{\link{logistic3}}
}

