% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mko.R
\name{modified.kostiakov}
\alias{modified.kostiakov}
\alias{modified.kostiakov.default}
\alias{predict.modified.kostiakov}
\title{Smith modified kostiakov infiltration parameter optmisation in R}
\usage{
modified.kostiakov(data = NULL, time, I, a = 0.1, k = 0.1, fc = 0.1,
  group = NULL)

\method{modified.kostiakov}{default}(data = NULL, time, I, a = 0.1,
  k = 0.1, fc = 0.1, group = NULL)

\method{predict}{modified.kostiakov}(object, time = NULL, ...)
}
\arguments{
\item{data}{dataframe. It can contain data with column names of "time" and "I"}

\item{time}{character or numeric. The name of time variable in the dataframe.
If the "data" parameter contains "time", this will be ignored.
The unit must be in seconds.}

\item{I}{cumulative infltration (I) or infiltration rate (i) [mm]}

\item{a}{constant}

\item{k}{constant}

\item{fc}{constant}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{object}{Model output object}

\item{...}{Any other graphical parameter}
}
\value{
\itemize{
\item{a:} {  optimised constant}
\item{k:} { optimised constant}
\item{fc}{optimised constant}
\item{output:} { output of the group simulation}
}
}
\description{
This function optimises Smith (1972) modified kostiakov
cumulative infltration (I)
}
\examples{
\dontrun{
data=read.csv(system.file("ext","sys","exampleBEST.csv",package="vadose"))
modified.kostiakov1<-modified.kostiakov(data=data,time="time",I="I")
#print(gof(modified.kostiakov1))
plot(modified.kostiakov1)
predict(modified.kostiakov1)
coef(modified.kostiakov1)
#gof(modified.kostiakov1)

#group simulation
data=read.csv(system.file("ext","sys","infiltration2.csv",package="vadose"))
assin_breko<- subset(data, ID=="41A20_1")
modified.kostiakov1vg<-modified.kostiakov(data=data,time="minutes",I="CumInfil",group="ID")
coef(modified.kostiakov1vg)
}
}
\author{
George Owusu
}
\references{
Smith, R. E. (1972). The infiltration envelope:
Results from a theoritical infiltrometer. J.Hydrol., 17, 1-21.
}

