% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OFEST.R
\name{OFEST}
\alias{OFEST}
\alias{OFEST.default}
\alias{cal.OFEST}
\alias{coef.OFEST}
\alias{gof.OFEST}
\alias{group.OFEST}
\alias{plot.OFEST}
\alias{predict.OFEST}
\title{The Offin Estimation of Soil Transfer parameters}
\usage{
OFEST(data = NULL, time, I, n = NULL, m = "b", pb = 1.2, tho = 0.169,
  thr = 0, ths = NULL, theta = NULL, PSD = NULL, Ks = NULL,
  h = seq(0, 1500, by = 5), model = "philip", type = "nonlinear",
  hg = "BEST", K = "BC")

\method{OFEST}{default}(data = NULL, time, I, n = NULL, m = "b",
  pb = 1.2, tho = 0.169, thr = 0, ths = NULL, theta = NULL,
  PSD = NULL, Ks = NULL, h = NULL, model = "philip",
  type = "nonlinear", hg = "BEST", K = "BC")

group.OFEST(data = NULL, time, I, n = NULL, m = "b", pb = 1.2,
  tho = 0.169, thr = 0, ths = NULL, theta = NULL, PSD = NULL,
  Ks = NULL, h = NULL, model = "all", type = "nonlinear", hg = "BEST",
  group, plot = TRUE, hlog = NULL, klog = NULL, layout = c(2, 2),
  opar = par(mar = c(2, 2, 1.8, 2)))

cal.OFEST(data = NULL, time, I, n = NULL, m = "b", pb = c(1, 1.8, 0.1),
  tho = c(0.01, 0.8, 0.05), thr = c(0, 0.2, 0.05), ths = NULL,
  theta = NULL, PSD = NULL, Ks = NULL, hg = "BEST", h = NULL,
  model = "all", type = "nonlinear", group = NULL)

gof.OFEST(x = NULL, obs = NULL, est = NULL, P = NULL)

\method{coef}{OFEST}(object, ...)

\method{predict}{OFEST}(object, h = NULL, ...)

\method{plot}{OFEST}(x, main = NULL, xlab = "Water Content", ylab = "",
  ylab2 = "", layout = NULL, hlog = NULL, klog = NULL,
  kcol = "darkgreen", col = c("blue", "red", "darkgreen", "gold"),
  units = c("s", "mm"), mfrow = c(2, 2), type = "all", legend = TRUE,
  opar = par(mar = c(2, 2, 1.5, 2)), ...)
}
\arguments{
\item{data}{dataframe. It can contain data with column names of "time" and "I"}

\item{time}{character or numeric. The name of time variable in the dataframe.
If the "data" parameter contains "time", this will be ignored.
The unit must be in seconds.}

\item{I}{character or numeric. The name of cumulative infiltration variable 
in the dataframe. If the "data" parameter contains "I", this will be ignored.
The unit must be in millimetres [mm].}

\item{n}{numeric. A shaping parameter for water retention curve. This can be calibrated.
If the parameter "PSD" is not set NULL \code{\link{lass3}} will be automatically used to estimate it.}

\item{m}{character. The water retention curve condition. 
It takes either "b" for Burdine condition or "m" for Mualem condition.}

\item{pb}{numeric. Bulk density[g cm^-3]. This can be calibrated.}

\item{tho}{numeric. initial volumetric soil water content. This can be calibrated.}

\item{thr}{numeric. Residual volumetric soil water content. This can be calibrated.}

\item{ths}{numeric. The saturated soil water content [m3/m3].}

\item{theta}{The  measured soil water content[m3m-3] at the modelled pressure levels.
It is used to check the fitness of the model.}

\item{PSD}{dataframe. Particle Size Distribution. A dataframe of two columns.
The first column should be labeled "D" in the range of 0.001 to 2 mm.
The second column should be named as "fr" i.e.' the fraction of Diameter. 
The range should be 0-1.}

\item{Ks}{saturated hydraulic conductivity}

\item{h}{list. The range of metric head desired or measured [mm].}

\item{model}{character. Either "philip" or "sw" (Swartzendruber) or "valiantzas" or "brutsaert".
The model for estimation of sorptivity (S) and hydraulic conductivity (Ks).}

\item{type}{character. Whether "linear" or "nonlinear" 
philip based equation to be applied.}

\item{hg}{character "rawls" or "BEST" bubbling capillary pressure  algorithm. 
It takes either Rawls (1993) as in Dingman(2002) or BEST}

\item{K}{character. The type of hydraulic conductivity model. 
It takes "BC" for Brooks and Corey (1964) and VG for van Genuchten (1980).}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{plot}{whether a plot should be performed.}

\item{hlog}{TRUE or FALSE. Whether metric potential should be log transformed}

\item{klog}{TRUE or FALSE. Whether hydraulic conductivity should be log transformed}

\item{layout}{plot layout}

\item{x}{a return object of the function.}

\item{obs}{Observed data}

\item{est}{Estimated or modelled data}

\item{P}{Number of Parameters}

\item{object}{Model output object}

\item{...}{Any other graphical parameter}

\item{main}{Title of the plot}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{ylab2}{y label of the second plot (K)}

\item{kcol}{colour of hydraulic conductivity}

\item{col}{Color of the plot}

\item{units}{Units of the plot}

\item{mfrow}{The graphical layout of the plots}

\item{legend}{A legend of the plot}
}
\value{
For the output of PSD see \code{\link{lass3}}.
For the output of goodness of fit tests see \code{\link{gof.OFEST}}. 
In the case of \code{cal.OFEST} the output can be assesed with mod object.
The return parameters can be assessed with \code{coef.OFEST}. 
The predicted soil water can be assessed  with \code{predict.OFEST} 
\itemize{
\item{Ks} { saturated hydraulic conductivity [LT^-1] }
\item{S} { Sorptivity  [LT^-0.5]}
\item{mod_theta,mod_thetal,mod_thetay,mod_thetab:} { The estimated soil water content at 
different metric potentials}
\item{theta:} { The observed soil water content at different metric potentials  }
\item{K:} { The unsaturated hydraulic conductivity [LT^-1] }
\item{hg:} { bubbling capillary pressure  [L] }
\item{q:} { infiltration rates [LT^-1] }
}
}
\description{
This function uses \code{\link{philip}} or \code{\link{sw}} 
with \code{\link{lass3}}, \code{\link{ksat}} and \code{\link{vg}} to estimate 
water retention and hydraulic conductivity curves. See datails below.
}
\details{
The function first estimates sorptivity and Ks from 3D or early part of 
vertical cumulative infiltration data with \code{\link{philip}} or \code{\link{sw}}
 or \code{\link{valiantzas}} or \code{\link{brutsaert}}.
Second, \code{\link{lass3}} is used to estimate pore-size parameters, e.g. n, b. Mualem or
Burdine condition is applied to estimate m parameter of van Genuchten hydraulic conductivity 
curve. Third, a bubbling capillary pressure hb is estimated with Ks and S. 
Fourth, van Genuchten (1980) water retention curve and hydraulic conductivity 
are estimated. Finally, the predicted infiltration rate (q) is estimated.
}
\examples{
psd=read.csv(system.file("ext","sys","psd.csv",package="vadose"))
hf=read.csv(system.file("ext","sys","h.csv",package="vadose"))
data=read.csv(system.file("ext","sys","exampleBEST.csv",package="vadose"))
ofin1<-OFEST(data=data,time="time",I="I",h=hf$h,PSD=psd)
pred=predict(ofin1)
coefficients=coef(ofin1)
plot(ofin1,type="h")
plot(ofin1,type="all")
plot(ofin1,type="all",hlog=TRUE,klog=TRUE,kcol="green")

#with measured theta
ofin2<-OFEST(data=data,time="time",I="I",h=hf$h,PSD=psd,model="sw",theta=ofin1$mod_theta)
print(gof.OFEST(ofin2))
ofin1$K
\dontrun{
#calibration
ofin2.cal<-cal.OFEST(data=data,time="time",I="I",h=hf$h,PSD=psd,model="sw",theta=ofin1$mod_theta)
coef2=coef(ofin2.cal)
predict=predict(ofin2.cal,h=500)
 }
}
\author{
George Owusu
}
\references{
\itemize{
\item{}{Swartzendruber, D. (1987). A quasi-solution of Richards equation 
for the downward infiltration of water into soil. Water Resour Res, 23, 809-817.}
\item{}{Philip, J. R. (1957). The theory of infiltration:Sorptivity and 
algebraic infiltration equations. Soil Science, 84, 257-264.}
\item{}{Brutsaert, W. (1977). Vertical infiltration in dry soil. 
Water Resour. Res., 13, 363-368.}
\item{}{Valiantzas, J. D. (2010). New linearized two-parameter infiltration equation for 
direct determination of conductivity and sorptivity. Journal of Hydrology, 387. 
doi: 10.1016/j.jhydrol.2009.12.049}
\item{}{Rawls, W. J., Ahuja, L. R., Brakensiek, D. L., & Shirmohammadi, A. (1993). 
Infiltration and Soil Water Movement. 
In D. Maidment (Ed.), Handbook of Hydrology: McGraw-Hill Education.}
\item{}{Dingman, S. L. (2002). Physical Hydrology: Prentice-Hall.}
\item{}{van Genuchten, M. T. (1980). A Closed-form Equation for Predicting the
 Hydraulic Conductivity of Unsaturated Soils1. Soil Sci. Soc. Am. J., 44(5), 892-898. 
 doi: 10.2136/sssaj1980.03615995004400050002x}
 \item{}{Brooks, R. H., & Corey, A. T. (1964). Hydraulic properties of porous 
 medium Hydrology Paper (Vol. 3): Colorado State University.}
}
}
\seealso{
\code{\link{BEST}}
}

