% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/offinbest.R
\docType{data}
\name{offinbest}
\alias{offinbest}
\title{Horinzontal Infiltration data of Offin Basin}
\format{A data frame with 302 rows and 9 variables:
\describe{
  \item{ID}{Unique ID of the data}
  \item{TownName}{Town Name}
  \item{time}{Recorded cumulative time [seconds]}
  \item{I}{Cumulative Infiltration [mm]}
  \item{pb}{Bulk density [g/cm3]}
  \item{r}{Radius of the cylinder ring}
  \item{tho}{initial soil water content [-]}
  \item{thr}{residual soil water content [-]}
  \item{n}{van Genuchten n parameter}
}}
\description{
A dataset containing horinzontal infiltration data of Offin Basin.
}
\author{
George Owusu
}

