% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offinbest.R
\docType{data}
\name{offinmeta}
\alias{offinmeta}
\title{Metadata of Infiltration data of Offin Basin}
\format{A data frame with 15 rows and 18 variables:
\describe{
  \item{ID}{Unique ID of the data}
  \item{TownName}{Town Name}
  \item{longitude}{longitudes in degrees}
  \item{latitude}{latitudes in degrees}
  \item{Time}{Total time of infiltration in minutes }
  \item{N}{the Number of water volumes (cylinders)}
  \item{q}{the mean infiltration rate in mm/s}
  \item{Texture}{USDA soil texture}
  \item{tho}{Initial soil water content (cm3/cm3)}
  \item{thr}{residual water content(cm3/cm3)}
  \item{pb}{bulk density (g/cm3)}
  \item{n}{van Genuchten n parameter}
  \item{m}{van Genuchten m parameter}
  \item{n2}{Particle Size parameter}
  \item{b}{Particle Size distribution parameter}
}}
\description{
A metadata  of  infiltration data of Offin Basin.
}
\author{
George Owusu
}

