% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/offinbest.R
\docType{data}
\name{offinpsd}
\alias{offinpsd}
\title{Particle Size Distribution Size of Offin Basin}
\format{A data frame with 311 rows and 4 variables:
\describe{
  \item{ID}{Uniquet ID of the data}
  \item{TownName}{Town Name}
  \item{D}{Particle diameter [mm]}
  \item{FractionSand}{proportion of particle sizes [0-1]}
  \item{fr}{proportion of particle sizes [0-1]}
}}
\description{
A dataset containing diameter and fractional proportion of soil particle in OFFin Basin
}
\author{
George Owusu
}

