% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/philip3.R
\name{philip}
\alias{coef.philip}
\alias{philip}
\alias{philip.default}
\alias{plot.philip}
\alias{predict.philip}
\alias{print.philip}
\alias{summary.philip}
\title{R optimisation of philip infiltration parameters}
\usage{
philip(data = NULL, time, I, S = 0.1, A = 0.1, type = "nonlinear",
  group = NULL)

\method{philip}{default}(data = NULL, time, I, S = 0.1, A = 0.1,
  type = "nonlinear", group = NULL)

\method{predict}{philip}(object, time = NULL, ...)

\method{plot}{philip}(x, xlab = "Time(Minutes)", ylab = "Cumulative (mm)",
  main = NULL, layout = NULL, ...)

\method{summary}{philip}(object, ...)

\method{print}{philip}(x, ...)

\method{coef}{philip}(object, ...)
}
\arguments{
\item{data}{dataframe. It can contain data with column names of "time" and "I"}

\item{time}{character or numeric. The name of time variable in the dataframe.
If the "data" parameter contains "time", this will be ignored.
The unit must be in seconds.}

\item{I}{cumulative infiltration (I) or infiltration rate (i) [mm]}

\item{S}{numeric. sorptivity parameter}

\item{A}{numeric.It is related Hydraulic Conductivity parameter}

\item{type}{character. It takes "linear" or "nonlinear" Philip equation}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{object}{Model output object}

\item{...}{Any other graphical parameter}

\item{x}{a return object of the function.}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{main}{Title of the plot}

\item{layout}{plot layout}
}
\value{
\itemize{
\item{A:} { opitmised  A parameter}
\item{Ks:} {  opitmised  saturated hydraulic conductivity=0.5*A [L/T]}
\item{S:} { optimised sorptivity [LT^-0.5]}
\item{output:} { output of the group simulation}
}
}
\description{
This function can linearly or nonlinearly optimise philip
cumulative infiltration (I) or infiltration rate (i) parameters: A and S.
}
\examples{
data=read.csv(system.file("ext","sys","exampleBEST.csv",package="vadose"))
philip1<-philip(data=data,time="time",I="I")
print(gof(philip1))
plot(philip1)
predict(philip1)
coef(philip1)
gof(philip1)

#infiltration rate
data=read.csv(system.file("ext","sys","infiltration2.csv",package="vadose"))
assin_breko<- subset(data, ID=="41A20_1")
philipr<-philip(data=assin_breko,time="minutes",I="cm.hr")

#group simulation
philipg<-philip(data=data,time="minutes",I="cm.hr",group="ID")
coef(philipg)
modprediction<-philip(time=1,S=0.1,A=0.5)
}
\author{
George Owusu
}
\references{
Philip, J. R. (1957). The theory of infiltration:Sorptivity and algebraic 
infiltration equations. Soil Science, 84, 257-264.
}

