% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmk.R
\name{revised.modified.kostiakov}
\alias{coef.revised.modified.kostiakov}
\alias{plot.revised.modified.kostiakov}
\alias{predict.revised.modified.kostiakov}
\alias{revised.modified.kostiakov}
\alias{revised.modified.kostiakov.default}
\alias{summary.revised.modified.kostiakov}
\title{Parhi Revised modified kostiakov infiltration parameter optmisation in R}
\usage{
revised.modified.kostiakov(data = NULL, time, I, a1 = 0.1, a2 = 0.1,
  b1 = 0.1, b2 = 0.1, group = NULL)

\method{revised.modified.kostiakov}{default}(data = NULL, time, I, a1 = 0.1,
  a2 = 0.1, b1 = 0.1, b2 = 0.1, group = NULL)

\method{predict}{revised.modified.kostiakov}(object, time = NULL, ...)

\method{plot}{revised.modified.kostiakov}(x, xlab = "Time(Minutes)",
  ylab = "Cumulative (cm)", main = NULL, layout = NULL, ...)

\method{summary}{revised.modified.kostiakov}(object, ...)

\method{coef}{revised.modified.kostiakov}(object, ...)
}
\arguments{
\item{data}{dataframe. It can contain data with column names of "time" and "I"}

\item{time}{character or numeric. The name of time variable in the dataframe.
If the "data" parameter contains "time", this will be ignored.
The unit must be in seconds.}

\item{I}{cumulative infltration (I) or infiltration rate (i) [mm]}

\item{a1}{constant}

\item{a2}{constant}

\item{b1}{constant}

\item{b2}{constant}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{object}{Model output object}

\item{...}{Any other graphical parameter}

\item{x}{a return object of the function.}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{main}{Title of the plot}

\item{layout}{plot layout}
}
\value{
\itemize{
\item{a1:} {  optimised constant}
\item{a2:} { optimised constant}
\item{b1}{optimised constant}
\item{b2}{optimised constant}
\item{output:} { output of the group simulation}
}
}
\description{
This function optimises Parhi et al (2007) revised modified kostiakov
cumulative infltration (I)
}
\examples{
data=read.csv(system.file("ext","sys","exampleBEST.csv",package="vadose"))
revised.modified.kostiakov1<-revised.modified.kostiakov(data=data,time="time",I="I")
#print(gof(revised.modified.kostiakov1))
plot(revised.modified.kostiakov1)
predict(revised.modified.kostiakov1)
coef(revised.modified.kostiakov1)
#gof(revised.modified.kostiakov1)

#group simulation
data=read.csv(system.file("ext","sys","infiltration2.csv",package="vadose"))
assin_breko<- subset(data, ID=="41A20_1")
revised.modified.kostiakov1vg<-revised.modified.kostiakov(data=data,time="minutes",
I="CumInfil",group="ID")
coef(revised.modified.kostiakov1vg)
}
\author{
George Owusu
}
\references{
Parhi, P. K., Mishra, S. K., & Singh, R. (2007). A modification to Kostiakov and
modified Kostiakov  infiltration modles. Water Resour. Manage., 21, 1973-1989.
}

