% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SCS.R
\name{scs}
\alias{plot.scs}
\alias{predict.scs}
\alias{scs}
\alias{scs.default}
\alias{summary.scs}
\title{SCS infiltration parameter optmisation in R}
\usage{
scs(data, time, I, a = 0.1, b = 0.1, group = NULL)

\method{scs}{default}(data, time, I, a = 0.1, b = 0.1, group = NULL)

\method{predict}{scs}(object, time = NULL, ...)

\method{plot}{scs}(x, xlab = "Time(Minutes)", ylab = "Cumulative (cm)",
  main = NULL, layout = NULL, ...)

\method{summary}{scs}(object, ...)
}
\arguments{
\item{data}{dataframe. It can contain data with column names of "time" and "I"}

\item{time}{character or numeric. The name of time variable in the dataframe.
If the "data" parameter contains "time", this will be ignored.
The unit must be in seconds.}

\item{I}{cumulative infltration (I) or infiltration rate (i) [mm]}

\item{a}{constant}

\item{b}{constant}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{object}{Model output object}

\item{...}{Any other graphical parameter}

\item{x}{a return object of the function.}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{main}{Title of the plot}

\item{layout}{plot layout}
}
\value{
\itemize{
\item{a:} {  optimised constant}
\item{b:} { optimised constant}
\item{output:} { output of the group simulation}
}
}
\description{
This function optimises Jury et al (1991)
cumulative infltration (I) parameters
}
\examples{
data=read.csv(system.file("ext","sys","exampleBEST.csv",package="vadose"))
SCS1<-scs(data=data,time="time",I="I")
#print(gof(SCS1))
plot(SCS1)
predict(SCS1)
coef(SCS1)
#gof(SCS1)

#group simulation
data=read.csv(system.file("ext","sys","infiltration2.csv",package="vadose"))
assin_breko<- subset(data, ID=="41A20_1")
SCSg<-scs(data=data,time="minutes",I="CumInfil",group="ID")
coef(SCSg)
}
\author{
George Owusu
}
\references{
Jury, W. A., Gardner, W. R., & H., G. W. (Eds.). (1991).
Soil physics (5th ed.). New York (NY): John Wiley & Sons.
}

