% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stroosnijder.R
\name{stroosnijder}
\alias{coef.stroosnijder}
\alias{plot.stroosnijder}
\alias{predict.stroosnijder}
\alias{print.stroosnijder}
\alias{stroosnijder}
\alias{stroosnijder.default}
\alias{summary.stroosnijder}
\title{stroosnijder infiltration parameter optimisation in R}
\usage{
stroosnijder(data = NULL, time, I, S = 0.1, K = 0.1, para = 2,
  group = NULL)

\method{stroosnijder}{default}(data = NULL, time, I, S = 0.1, K = 0.1,
  para = 2, group = NULL)

\method{predict}{stroosnijder}(object, time = NULL, ...)

\method{plot}{stroosnijder}(x, xlab = "Time(Minutes)",
  ylab = "Cumulative (cm)", main = NULL, layout = NULL, ...)

\method{summary}{stroosnijder}(object, ...)

\method{print}{stroosnijder}(x, ...)

\method{coef}{stroosnijder}(object, ...)
}
\arguments{
\item{data}{dataframe. It can contain data with column names of "time" and "I"}

\item{time}{character or numeric. The name of time variable in the dataframe.
If the "data" parameter contains "time", this will be ignored.
The unit must be in seconds.}

\item{I}{cumulative infiltration (I) or infiltration rate (i) [mm]}

\item{S}{numeric. sorptivity parameter}

\item{K}{character. The type of hydraulic conductivity model. 
It takes "BC" for Brooks and Corey (1964) and VG for van Genuchten (1980).}

\item{para}{a parameter for each model}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{object}{Model output object}

\item{...}{Any other graphical parameter}

\item{x}{a return object of the function.}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{main}{Title of the plot}

\item{layout}{plot layout}
}
\value{
\itemize{
\item{Ks:} {  optimised saturated hydraulic conductivity [L/T]}
\item{S:} { optimised sorptivity [LT^-0.5]}
\item{output:} { output of the group simulation}
}
}
\description{
This function optimises stroosnijder (1976)
cumulative infiltration  (I): Ks and S.
}
\examples{
data=read.csv(system.file("ext","sys","exampleBEST.csv",package="vadose"))
stroosnijder1<-stroosnijder(data=data,time="time",I="I")
#print(gof(stroosnijder1))
plot(stroosnijder1)
predict(stroosnijder1)
coef(stroosnijder1)
#gof(stroosnijder1)

#group simulation
data=read.csv(system.file("ext","sys","infiltration2.csv",package="vadose"))
assin_breko<- subset(data, ID=="41A20_1")
stroosnijderg<-stroosnijder(data=data,time="minutes",I="CumInfil",group="ID")
coef(stroosnijderg)
}
\author{
George Owusu
}
\references{
stroosnijder, L., 1976. Infiltratie en Herverdeling van Water in Gronde. 
Versl.Landbouwkd, Onderz. 847
}

