% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SW.R
\name{sw}
\alias{coef.sw}
\alias{plot.sw}
\alias{predict.sw}
\alias{print.sw}
\alias{sw}
\alias{sw.default}
\title{Swartzendruber infiltration parameter optmisation in R}
\usage{
sw(data, time, I, As = 0.1, S = 0.1, Ks = 0.1, group = NULL)

\method{sw}{default}(data, time, I, As = 0.1, S = 0.1, Ks = 0.1,
  group = NULL)

\method{predict}{sw}(object, time = NULL, ...)

\method{plot}{sw}(x, xlab = "Time(Minutes)", ylab = "Cumulative (mm)",
  main = NULL, layout = NULL, ...)

\method{print}{sw}(x, ...)

\method{coef}{sw}(object, ...)
}
\arguments{
\item{data}{dataframe. It can contain data with column names of "time" and "I"}

\item{time}{character or numeric. The name of time variable in the dataframe.
If the "data" parameter contains "time", this will be ignored.
The unit must be in seconds.}

\item{I}{cumulative infltration (I) or infiltration rate (i) [mm]}

\item{As}{an optimised parameter}

\item{S}{numeric. sorptivity parameter}

\item{Ks}{saturated hydraulic conductivity}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{object}{Model output object}

\item{...}{Any other graphical parameter}

\item{x}{a return object of the function.}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{main}{Title of the plot}

\item{layout}{plot layout}
}
\value{
\itemize{
\item{As:} { optmised As parameter}
\item{Ks:} {  optmised saturated hydraulic conductivity[L/T]}
\item{S:} { optimised sorptivity [LT^-0.5]}
\item{output:} { output of the group simulation}
}
}
\description{
This function optmises Swartzendruber (1987)
cumulative infltration (I) or infltration rate (i) parameters: As, Ks and S.
}
\examples{
data=read.csv(system.file("ext","sys","exampleBEST.csv",package="vadose"))
sw1<-sw(data=data,time="time",I="I")
print(gof(sw1))
plot(sw1)
predict(sw1)
coef(sw1)
gof(sw1)

#infiltration rate
data=read.csv(system.file("ext","sys","infiltration2.csv",package="vadose"))
assin_breko<- subset(data, ID=="41A20_1")
swr<-sw(data=assin_breko,time="minutes",I="cm.hr")

#group simulation
swg<-sw(data=data,time="minutes",I="cm.hr",group="ID")
coef(swg)
}
\author{
George Owusu
}
\references{
Swartzendruber, D. (1987). A quasi-solution of Richards equation for
the downward infiltration of water into soil. Water Resour Res, 23, 809-817.
}

