% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lass3.R
\name{texture}
\alias{texture}
\title{Prediction of USDA Soil texture lassification from PSD object}
\usage{
texture(object = NULL, sand = NULL, silt = NULL, clay = NULL)
}
\arguments{
\item{object}{A PSD object from \code{\link{lass3}}}

\item{sand}{In case object is NULL, percentage of Sand can be supplied}

\item{silt}{In case object is NULL, percentage of Silt can be supplied}

\item{clay}{Incase object is NULL, percentage of Clay can be supplied}
}
\value{
texture
}
\description{
Prediction of USDA Soil texture lassification from PSD object
}
\examples{
file="C:/Users/GeoKings/Documents/George Owusu/UG PhD/Data/soil2.csv"
data=read.csv(file)
single<- subset(data, ID=="30B20_1")
mod=lass3 (data=single,p="sand",D="D",fr="Sand.")
texture(mod)
}
\author{
George Owusu
}

