% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vadose.rank.R
\name{vadose.rank}
\alias{vadose.rank}
\title{Ranking and  model performance assessment of vadose model object}
\usage{
vadose.rank(data, var = c("r2", "d", "E", "RMSE", "ARE", "MBE"),
  model = "model", fun = "median", xlab = NULL, ylab = c("R-square",
  "Index of Agreement", "Nash-Sutcliffe", "RMSE", "Average Relative Error",
  "Mean Absolute Error"), main = c("R-square", "Index of Agreement",
  "Nash-Sutcliffe", "RMSE", "Average Relative Error", "Mean Bias Error"),
  ylim = list(c(0.99, 1), c(0.98, 1), c(0.9, 1), c(0, 2.5), c(0, 30), c(-0.2,
  0.2)), las = 2, opar = par(mfrow = c(2, 3), mar = c(7, 5, 2, 1)),
  myrank = NULL)
}
\arguments{
\item{data}{dataframe of goodness of fit indicators}

\item{var}{list of the indicators or variables that should be used to rank the models}

\item{model}{character. The column name indicating the models that are to be ranked.}

\item{fun}{Either median or mean or max or min or any of the fun objects}

\item{xlab}{The x label of the boxplot}

\item{ylab}{The y label of the boxplot}

\item{main}{The title of the boxplot}

\item{ylim}{The y limits of the boxplot}

\item{las}{The orientation  of x labels}

\item{opar}{The par attributes}

\item{myrank}{a redundant variable for internal use}
}
\value{
rank and fun.summary indicating the aggregation of the data by fun
}
\description{
This function ranks and boxplots \code{\link{BEST}},  \code{\link{OFEST}} or
\code{\link{lass3}} goodness of fit objects. Various models in  \code{\link{BEST}} 
and  \code{\link{OFEST}} are ranked by either median or mean or max or any of the fun objects.
}
\examples{
data(offinbest) #infiltration data
data(offinpsd) #Particle size Distribution data
data(htheta) #measured h and theta
PSD=offinpsd
Figure2=group.BEST(data = offinbest, group="TownName",PSD=PSD,layout=c(3,4),hlog=TRUE,plot=FALSE)
Figure3=group.OFEST(data = offinbest, hg="best",group="TownName",PSD=PSD,layout=c(3,4),hlog=TRUE,type = "nonlinear",plot=FALSE)
data=rbind(Figure2$statistics2,Figure3$statistics2)
ranking=vadose.rank(data)
ranking$rank

MBE=vadose.rank(data,var="MBE",ylim=c(-1,2),ylab=NULL,main=NULL)
MPE=ranking=vadose.rank(data,var="MPE",ylim=c(-20,30),ylab=NULL,main=NULL)
}
\author{
George Owusu
}

