% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vadoseerror2.R
\name{vadose.tryCatch}
\alias{vadose.tryCatch}
\title{Internal function to Catch  errors and warnings}
\usage{
vadose.tryCatch(expr)
}
\arguments{
\item{expr}{an  R expression to evaluate}
}
\value{
value a list with 'value' and 'warning', where value may be an error caught
}
\description{
Internal function to Catch  errors and warnings
}
\examples{
thiserror=str( vadose.tryCatch( log( "a" ) ) )$value
}

