% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valiantzas.R
\name{valiantzas}
\alias{coef.valiantzas}
\alias{plot.valiantzas}
\alias{predict.valiantzas}
\alias{print.valiantzas}
\alias{summary.valiantzas}
\alias{valiantzas}
\alias{valiantzas.default}
\title{Valiantzas infiltration parameter optimization in R}
\usage{
valiantzas(data = NULL, time, I, S = 0.1, Ks = 0.3, type = "linear",
  ratio = 0.5, group = NULL)

\method{valiantzas}{default}(data = NULL, time, I, S = 0.1, Ks = 0.3,
  type = "linear", ratio = 0.5, group = NULL)

\method{predict}{valiantzas}(object, time = NULL, ...)

\method{plot}{valiantzas}(x, xlab = "Time(Minutes)",
  ylab = "Cumulative (mm)", main = NULL, layout = NULL, ...)

\method{summary}{valiantzas}(object, ...)

\method{print}{valiantzas}(x, ...)

\method{coef}{valiantzas}(object, ...)
}
\arguments{
\item{data}{dataframe. It can contain data with column names of "time" and "I"}

\item{time}{character or numeric. The name of time variable in the dataframe.
If the "data" parameter contains "time", this will be ignored.
The unit must be in seconds.}

\item{I}{cumulative infiltration (I) or infiltration rate (i) [mm]}

\item{S}{numeric. sorptivity parameter}

\item{Ks}{saturated hydraulic conductivity}

\item{type}{character. It takes "linear" or "nonlinear" Philip equation}

\item{ratio}{a parameter expressing relationship between philip A and Ks}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{object}{Model output object}

\item{...}{Any other graphical parameter}

\item{x}{a return object of the function.}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{main}{Title of the plot}

\item{layout}{plot layout}
}
\value{
\itemize{
\item{Ks:} {  optimised saturated hydraulic conductivity}
\item{S:} { optimised sorptivity [LT^-0.5]}
\item{ratio:} { optimised ratio parameter}
\item{output:} { output of the group simulation}
}
}
\description{
This function optimises Valiantzas (2010)
cumulative infiltration (I)  parameters: Ks, S and ratio between philip A and Ks.
}
\examples{
data=read.csv(system.file("ext","sys","exampleBEST.csv",package="vadose"))
Valiantzas1<-valiantzas(data=data,time="time",I="I")
plot(Valiantzas1)
predict(Valiantzas1)
coef(Valiantzas1)


#infiltration rate
data=read.csv(system.file("ext","sys","infiltration2.csv",package="vadose"))
assin_breko<- subset(data, ID=="41A20_1")
Valiantzasr<-valiantzas(data=assin_breko,time="minutes",I="cm.hr")

#group simulation
Valiantzasg<-valiantzas(data=data,time="minutes",I="cm.hr",group="ID")
coef(Valiantzasg)
#generic function
}
\author{
George Owusu
}
\references{
Valiantzas, J. D. (2010). New linearized two-parameter infiltration equation 
for direct determination of conductivity and sorptivity. 
Journal of Hydrology, 387. doi: 10.1016/j.jhydrol.2009.12.049
}

