% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VG.R
\name{vg}
\alias{coef.vg}
\alias{plot.vg}
\alias{predict.vg}
\alias{print.vg}
\alias{summary.vg}
\alias{vg}
\alias{vg.default}
\title{R estimation of van Genuchten water retention curve and hydraulic conductivity with intenal Ks}
\usage{
vg(data = NULL, h, theta = NULL, thr = 0.1, ths = 0.1, alp = 0.1,
  n = 1, m = "m", Ks, para = NULL, group = NULL)

\method{vg}{default}(data = NULL, h, theta = NULL, thr = 0.1, ths = 0.1,
  alp = 0.1, n = 1, m = "m", Ks = "Nasta", para = NULL,
  group = NULL)

\method{predict}{vg}(object, h = NULL, ...)

\method{plot}{vg}(x, main = NULL, xlab = "Water Content",
  ylab = "Tension Head,h(cm)", ylab2 = "Hydraulic Conductivity, K (cm/day)",
  layout = NULL, kcol = "green", hcol = "black", hlog = NULL,
  klog = NULL, ...)

\method{summary}{vg}(object, ...)

\method{print}{vg}(x, ...)

\method{coef}{vg}(object, ...)
}
\arguments{
\item{data}{dataframe that contains pressure head "h" and the corresponding water content "theta"}

\item{h}{character. The name of the pressure head as in the data parameter above.}

\item{theta}{character. The name of the water content as in the data parameter above}

\item{thr}{numeric. Residual volumetric soil water content. This can be calibrated.}

\item{ths}{numeric. The saturated soil water content [m3/m3].}

\item{alp}{van Genuchten water retention alpha parameter}

\item{n}{numeric. A shaping parameter for water retention curve. This can be calibrated.
If the parameter "PSD" is not set NULL \code{\link{lass3}} will be automatically used to estimate it.}

\item{m}{character. The water retention curve condition.
It takes either "b" for Burdine condition or "m" for Mualem condition.}

\item{Ks}{saturated hydraulic conductivity}

\item{para}{character. This is useful if the function is used for
estimation instead of optimisation. It can be combined with Ks parameter to retrieve
hydraulic parameters from Carsel&Parrish (1988). If para is set to "soil" and Ks is set to
"Clay", the parameters of clay soil from Carsel&Parrish (1988) will be used for estimation.
see example below.}

\item{group}{character. The name of the group variables if the data is from different areas.}

\item{object}{Model output object}

\item{...}{Any other graphical parameter}

\item{x}{a return object of the function.}

\item{main}{Title of the plot}

\item{xlab}{x label of the plot}

\item{ylab}{y label of the plot}

\item{ylab2}{y label of the second plot (K)}

\item{layout}{plot layout}

\item{kcol}{colour of hydraulic conductivity}

\item{hcol}{colour of pressure head}

\item{hlog}{TRUE or FALSE. Whether metric potential should be log transformed}

\item{klog}{TRUE or FALSE. Whether hydraulic conductivity should be log transformed}
}
\value{
optimised parameters
\itemize{
\item{Ks} { saturated hydraulic conductivity [LT^-1] }
\item{mod_theta,mod_thetal,mod_thetay,mod_thetab:} { The estimated soil water content at
different metric potentionals}
\item{theta:} { The observed soil water content at different metric potentionals  }
\item{K:} { The unsaturated hydraulic conductivity [LT^-1] }
\item{se:} { effective saturatn}
\item{thr} { optmised residual water content[LTLT^-1] }
\item{ths} { optmised saturated water content[LTLT^-1] }
\item{alp} { optmised alpha value}
\item{n} {optmised n shaping parameter }
\item{Kr} { The ratio of K and Ks }
}
}
\description{
This function estimate and optimise van Genuchten water rentention and hydraulic conductivity.
}
\examples{
\dontrun{
 #optimisation with single data
datartc=read.csv(system.file("ext","sys","retentionVG.csv",package="vadose"))
modrtc<-vg(data=datartc,h="h",theta="theta",thr=0.1, ths=0.1, alp=0.1, n=1)
plot(modrtc)

#optimisation with group data isric

data=read.csv(system.file("ext","sys","isric2.csv",package="vadose"))
#optimisation with group data
#used public initials and multiple Ks
modisrc<-vg(data=data,h="x",theta="y",m="b",thr=0.1, ths=0.1, alp=0.1, n=1,group="Sample",
Ks=c("Sand","Clay","Silt","silty clay loam"),para="soil")
modisrc<-vg(data=data,h="x",theta="y",m="b",thr=0.1, ths=0.3, alp=0.01, n=2,group="Sample")
plot(modisrc)
mod=vg(h=200)
}
}
\author{
George Owusu
}
\references{
\itemize{
\item{}{van Genuchten, M. T. (1980). A Closed-form Equation for Predicting the
 Hydraulic Conductivity of Unsaturated Soils1. Soil Sci. Soc. Am. J., 44(5), 892-898.
 doi: 10.2136/sssaj1980.03615995004400050002x}
\item{}{Carsel, R. F., & Parrish, R. S. (1988). Developing joint probability distributions of
soil water retention characteristics. Water Resour. Res., 24, 755-769.}
}
}

